[MOD] 기록 추가, userId별로 띄우는거 구현

+++ src/main/java/com/example/letsGo/dao/RecordDao.java
@@ -8,8 +8,8 @@ import java.util.List;
 
 public interface RecordDao {
     Record addRecord(Record record);
-    Record updateRecord(int recordId, Record record);
-    void deleteRecord(int recordId);
-    Record getRecord(int recordId);
+    Record updateRecord(Long recordId, Record record);
+    void deleteRecord(Long recordId);
+    Record getRecord(Long recordId);
     List<Record> getAllRecords(int userId);
 }
