+++ src/main/java/com/example/letsGo/dao/mybatis/mapper/RecordMapper.java
@@ -8,8 +8,8 @@ import java.util.List;
 @Mapper
 public interface RecordMapper {
     Record addRecord(Record record);
-    Record updateRecord(int recordId, Record record);
-    void deleteRecord(int recordId);
-    Record getRecord(int recordId);
+    Record updateRecord(Long recordId, Record record);
+    void deleteRecord(Long recordId);
+    Record getRecord(Long recordId);
     List<Record> getAllRecords(int userId);
 }
