+++ src/main/java/com/example/letsGo/service/RecordService.java
@@ -7,9 +7,9 @@ import java.util.List;
 
 @Service
 public interface RecordService {
-    Record getRecord(int recordId);
+    Record getRecord(Long recordId);
     List<Record> getAllRecords(int userId);
     Record addRecord(Record record);
-    Record updateRecord(int recordId, Record record);
-    void deleteRecord(int recordId);
+    Record updateRecord(Long recordId, Record record);
+    void deleteRecord(Long recordId);
 }
